Update to latest versions of dependencies

+++ src/core/store/io/mmap_index_input.rs
@@ -176,7 +176,7 @@ impl MmapIndexInput {
             };
             let mmap = unsafe {
                 MmapOptions::new()
-                    .offset(offset)
+                    .offset(offset as u64)
                     .len(adapted_len)
                     .map(&file)?
             };
